requirement.txt
beautifulsoup4==4.4.1
DistributedLock==1.2
Django==1.8.5
django-annoying==0.8.4
django-m2m-history==0.3.2
django-oauth-tokens==0.5.8
django-picklefield==0.3.2
django-taggit==0.17.3
django-tinymce==2.0.5
django-twitter-api==0.2.6
MySQL-python==1.2.5
oauthlib==1.0.3
python-memcached==1.57
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy==3.4.0
